+++ sdccc/src/main/java/com/draeger/medical/sdccc/messages/MessageStorage.java
@@ -54,6 +54,7 @@ import java.util.concurrent.BrokenBarrierException;
 import java.util.concurrent.CyclicBarrier;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.locks.ReentrantLock;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -171,7 +172,7 @@ public class MessageStorage implements AutoCloseable {
 
     private final TestRunObserver testRunObserver;
     private final boolean summarizeMessageEncodingErrors;
-    private long messageEncodingErrorCount;
+    private AtomicLong messageEncodingErrorCount;
     private int invalidMimeTypeCount;
     private final boolean enableEncodingCheck;
 
@@ -189,7 +190,7 @@ public class MessageStorage implements AutoCloseable {
         this.blockingQueueSize = blockingQueueSize;
         this.summarizeMessageEncodingErrors = summarizeMessageEncodingErrors;
         this.enableEncodingCheck = enableEncodingCheck;
-        this.messageEncodingErrorCount = 0;
+        this.messageEncodingErrorCount = new AtomicLong(0);
         this.invalidMimeTypeCount = 0;
 
         this.actionExtractor = new XPathExtractor(String.format("//%s:Action", WsAddressingConstants.NAMESPACE_PREFIX));
@@ -284,7 +285,7 @@ public class MessageStorage implements AutoCloseable {
                 if (this.enableEncodingCheck) {
                     if (this.summarizeMessageEncodingErrors) {
                         // TestRun will be invalidated in TestSuite if messageEncodingErrorCount > 0
-                        this.messageEncodingErrorCount += 1;
+                        this.messageEncodingErrorCount.incrementAndGet();
                     } else {
                         this.testRunObserver.invalidateTestRun(String.format(
                                 "Encountered message encoding problem: charset %s was specified, but message "
@@ -374,7 +375,7 @@ public class MessageStorage implements AutoCloseable {
                     (String originA, String valueA, String originB, String valueB) -> {
                         if (this.summarizeMessageEncodingErrors) {
                             // TestRun will be invalidated in TestSuite if messageEncodingErrorCount > 0
-                            this.messageEncodingErrorCount += 1;
+                            this.messageEncodingErrorCount.incrementAndGet();
                         } else {
                             this.testRunObserver.invalidateTestRun(String.format(
                                     INCONSISTENT_CHARSET_DECLARATION_WITH_ORIGINS,
@@ -398,7 +399,7 @@ public class MessageStorage implements AutoCloseable {
                 }
                 if (this.summarizeMessageEncodingErrors) {
                     // TestRun will be invalidated in TestSuite if messageEncodingErrorCount > 0
-                    this.messageEncodingErrorCount++;
+                    this.messageEncodingErrorCount.incrementAndGet();
                 } else {
                     this.testRunObserver.invalidateTestRun(String.format(
                             "Encountered a message whose encoding is declared to be %s. This violates"
@@ -433,7 +434,7 @@ public class MessageStorage implements AutoCloseable {
                 } else {
                     if (this.summarizeMessageEncodingErrors) {
                         // TestRun will be invalidated in TestSuite if messageEncodingErrorCount > 0
-                        this.messageEncodingErrorCount += 1;
+                        this.messageEncodingErrorCount.incrementAndGet();
                     } else {
                         this.testRunObserver.invalidateTestRun(
                                 "Message encoding could not be determined for message with ID '" + message.getID()
@@ -541,7 +542,7 @@ public class MessageStorage implements AutoCloseable {
         } catch (IOException e) {
             if (this.summarizeMessageEncodingErrors) {
                 // TestRun will be invalidated in TestSuite if messageEncodingErrorCount > 0
-                this.messageEncodingErrorCount += 1;
+                this.messageEncodingErrorCount.incrementAndGet();
             } else {
                 this.testRunObserver.invalidateTestRun(
                         "Unable to read message contents of message with ID '" + message.getID() + "'", e);
@@ -570,7 +571,7 @@ public class MessageStorage implements AutoCloseable {
                 (String originA, String valueA, String originB, String valueB) -> {
                     if (this.summarizeMessageEncodingErrors) {
                         // TestRun will be invalidated in TestSuite if messageEncodingErrorCount > 0
-                        this.messageEncodingErrorCount += 1;
+                        this.messageEncodingErrorCount.incrementAndGet();
                     } else {
                         this.testRunObserver.invalidateTestRun(String.format(
                                 INCONSISTENT_CHARSET_DECLARATION_WITH_ORIGINS,
@@ -736,7 +737,7 @@ public class MessageStorage implements AutoCloseable {
                         //       the Message before storing it in the DB.
                         if (this.summarizeMessageEncodingErrors) {
                             // TestRun will be invalidated in TestSuite if messageEncodingErrorCount > 0
-                            this.messageEncodingErrorCount += 1;
+                            this.messageEncodingErrorCount.incrementAndGet();
                         } else {
                             this.testRunObserver.invalidateTestRun(
                                     String.format(
@@ -1820,7 +1821,7 @@ public class MessageStorage implements AutoCloseable {
      * @return the count
      */
     public long getMessageEncodingErrorCount() {
-        return this.messageEncodingErrorCount;
+        return this.messageEncodingErrorCount.getPlain();
     }
 
     /**
@@ -1832,7 +1833,7 @@ public class MessageStorage implements AutoCloseable {
     }
 
     /**
-     * Container for the query result stream and the information on whether or not the objects are present. This shall
+     * Container for the query result stream and the information on whether the objects are present. This shall
      * always be closed after usage!
      *
      * @param <T> query result stream type