+++ CHANGELOG.md
@@ -14,8 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - message encoding errors are summarized by default, but all information is available via the configuration option
   SDCcc.SummarizeMessageEncodingErrors.
 - checking for message encoding errors can be disabled using the configuration option SDCcc.EnableMessageEncodingCheck.
+  - when enabled, SDCcc also checks if message bodies can be decoded using the encoding specified.
 - configuration option SDCcc.Network.MulticastTTL to configure the Time To Live of the Multicast Packets used for Discovery.
-- SDCcc now checks the encodings of all messages when SDCcc.EnableMessageEncodingCheck is enabled.
 
 ### Changed
 - SDCri version 4.1.0-SNAPSHOT